requirements.txt
# Core Data Processing
pandas>=1.5.0
numpy>=1.24.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0
imbalanced-learn>=0.11.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Web Application
streamlit>=1.28.0

# Utilities
pyyaml>=6.0
joblib>=1.3.0

# Development & Testing
pytest>=7.4.0
pytest-cov>=4.1.0
jupyter>=1.0.0
